move abspath to file pkg.
