+++ osmops/cfg/store.go
@@ -8,13 +8,14 @@ import (
 	"strings"
 
 	u "github.com/fluxcd/source-watcher/osmops/util"
+	"github.com/fluxcd/source-watcher/osmops/util/file"
 )
 
 // Store holds the OSM Ops program configuration read from the OSM Ops
 // config and credentials files.
 type Store struct {
-	rootDir   u.AbsPath
-	targetDir u.AbsPath
+	rootDir   file.AbsPath
+	targetDir file.AbsPath
 	fileExt   []u.NonEmptyStr
 	osmCreds  *OsmConnection
 }
@@ -23,7 +24,7 @@ type Store struct {
 // their content, and packs the content in a Store. If an I/O error happens
 // when reading the files or some of the YAML content isn't valid, NewStore
 // returns a error. Each YAML type documents what a valid instance is.
-func NewStore(repoRootDir u.AbsPath) (*Store, error) {
+func NewStore(repoRootDir file.AbsPath) (*Store, error) {
 	var err error
 	var cfg *OpsConfig
 	s := Store{rootDir: repoRootDir}
@@ -53,7 +54,7 @@ func NewStore(repoRootDir u.AbsPath) (*Store, error) {
 // repo root directory.
 const OpsConfigFileName = "osm_ops_config.yaml"
 
-func readConfig(rootDir u.AbsPath) (*OpsConfig, error) {
+func readConfig(rootDir file.AbsPath) (*OpsConfig, error) {
 	file := rootDir.Join(OpsConfigFileName)
 	if fileData, err := ioutil.ReadFile(file.Value()); err != nil {
 		return nil, err
@@ -62,7 +63,7 @@ func readConfig(rootDir u.AbsPath) (*OpsConfig, error) {
 	}
 }
 
-func buildTargetDirPath(rootDir u.AbsPath, cfg *OpsConfig) (u.AbsPath, error) {
+func buildTargetDirPath(rootDir file.AbsPath, cfg *OpsConfig) (file.AbsPath, error) {
 	target := rootDir.Join(cfg.TargetDir)
 	if err := target.IsDir(); err != nil {
 		return target, err
@@ -70,14 +71,14 @@ func buildTargetDirPath(rootDir u.AbsPath, cfg *OpsConfig) (u.AbsPath, error) {
 	return target, nil
 }
 
-func buildCredsDirPath(rootDir u.AbsPath, cfg *OpsConfig) (u.AbsPath, error) {
+func buildCredsDirPath(rootDir file.AbsPath, cfg *OpsConfig) (file.AbsPath, error) {
 	if filepath.IsAbs(cfg.ConnectionFile) {
-		return u.ParseAbsPath(cfg.ConnectionFile)
+		return file.ParseAbsPath(cfg.ConnectionFile)
 	}
 	return rootDir.Join(cfg.ConnectionFile), nil
 }
 
-func readCreds(rootDir u.AbsPath, cfg *OpsConfig) (*OsmConnection, error) {
+func readCreds(rootDir file.AbsPath, cfg *OpsConfig) (*OsmConnection, error) {
 	var fileData []byte
 	if credsFile, err := buildCredsDirPath(rootDir, cfg); err != nil {
 		return nil, err
@@ -112,13 +113,13 @@ func getFileExtensions(cfg *OpsConfig) []u.NonEmptyStr {
 }
 
 // RepoRootDirectory returns the absolute path to the repo root directory.
-func (s *Store) RepoRootDirectory() u.AbsPath {
+func (s *Store) RepoRootDirectory() file.AbsPath {
 	return s.rootDir
 }
 
 // RepoTargetDirectory returns the absolute path to the directory within
 // the repo where to find OSM Git Ops files.
-func (s *Store) RepoTargetDirectory() u.AbsPath {
+func (s *Store) RepoTargetDirectory() file.AbsPath {
 	return s.targetDir
 }
