+++ osmops/cfg/store_test.go
@@ -8,15 +8,16 @@ import (
 	"testing"
 
 	u "github.com/fluxcd/source-watcher/osmops/util"
+	"github.com/fluxcd/source-watcher/osmops/util/file"
 )
 
-func findTestDataDir(dirIndex int) u.AbsPath {
+func findTestDataDir(dirIndex int) file.AbsPath {
 	_, thisFileName, _, _ := runtime.Caller(1)
 	enclosingDir := filepath.Dir(thisFileName)
 	testDataDirName := fmt.Sprintf("test_%d", dirIndex)
 	testDataDir := filepath.Join(enclosingDir, "store_test_dir",
 		testDataDirName)
-	p, _ := u.ParseAbsPath(testDataDir)
+	p, _ := file.ParseAbsPath(testDataDir)
 
 	return p
 }
